Moved 'get or create metric' logic to MetricServiceImpl from MetricManager.
